+++ cloverExtension/src/components/ImageTopic.tsx
@@ -7,14 +7,11 @@ await loadCv()
 const cv = loadCv()
 
 export default function ImageTopic(props: { topic: string }) {
-  const message = useTopic(props.topic)
+  const message = useTopic(props.topic, true)
   const canvas = useRef<HTMLCanvasElement>(null)
   useEffect(() => {
     if (message != null) {
-      const arr = Uint8Array.from(atob(message.data as string), (c) =>
-        c.charCodeAt(0)
-      )
-      console.log(message.encoding)
+      const arr = message.data
       const mat = cv.matFromArray(
         message.height,
         message.width,
@@ -27,6 +24,7 @@ export default function ImageTopic(props: { topic: string }) {
         cv[`COLOR_${getCvConversion(message.encoding as string)}2RGB`]
       )
       cv.imshow(canvas.current, mat)
+      mat.delete()
     }
   }, [message?.data])