+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 - (cloverExtension) package.json теперь генирируется автоматически из файла tools.ts
 - (cloverExtension) Порефакторили конфигурацию инструментов
 - gitignore теперь игнорирует yarn-error.log и \_\_pycache\_\_
+- (cloverExtension) Использование CBOR в image viewer
+- (cloverExtension) Осовбождение памяти в image viewer
+- (cloverExtension) Добавлени компонент для просмотра информации о полёте (как в QGroundControl)